+++ android/core/engine/src/test/java/com/xianxia/sect/core/engine/domain/exploration/DiscipleDeathHandlerTest.kt
@@ -3,11 +3,14 @@ package com.xianxia.sect.core.engine.domain.exploration
 import com.xianxia.sect.core.exploration.DiscipleDeathHandler
 import com.xianxia.sect.core.model.DiscipleStatus
 import com.xianxia.sect.core.state.DiscipleTables
+import com.xianxia.sect.core.state.WriteGuardRule
 import org.junit.Assert.*
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
 
 class DiscipleDeathHandlerTest {
+    @get:Rule val writeGuardRule = WriteGuardRule()
 
     private lateinit var handler: DiscipleDeathHandler
     private lateinit var tables: DiscipleTables